Extracted abstract class with common methods for both algorithms.
