--- tasks/andrii.lysenko/homework_1/ITextParser.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace text_spacing
-{
-    public interface ITextParser
-    {
-        void ReadDictionary(string path);
-        void SetDictionary(string[] dictionary);
-        IEnumerable<string> SplitText(string textWithoutSpaces);
-    }
-}
